CI/Tests/Workflow: Build-Fixes, Push-only Workflow, User/License Plan
- UpdateChecker: intptr_t für ShellExecuteA (Windows C4311)
- Tests: SimplifyService, VisualizationService, RoutingService öffentliche API
- EndToEndTest: Sketch.addLine(Point2D), fstream, exportAssemblyToStep
- UpdateServiceTest: CurrentVersion statt privater Checksum-Methoden
- cad_interop verlinkt cad_core; viewport_integration_test verlinkt cad_core
- Workflow: Nur Push (main/master), kein GitHub-Release mehr
- docs: PLAN_USER_AND_LICENSING.md, TODO_USER_BACKEND_LICENSES.md

+++ src/core/updates/UpdateChecker.cpp
@@ -1,6 +1,7 @@
 #include "UpdateChecker.h"
 
 #include <array>
+#include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <sstream>
@@ -178,7 +179,7 @@ bool openUrlInBrowser(const std::string& url) {
     if (url.empty()) {
         return false;
     }
-    return reinterpret_cast<int>(ShellExecuteA(nullptr, "open", url.c_str(), nullptr, nullptr, SW_SHOWNORMAL)) > 32;
+    return reinterpret_cast<intptr_t>(ShellExecuteA(nullptr, "open", url.c_str(), nullptr, nullptr, SW_SHOWNORMAL)) > 32;
 #elif __APPLE__
     if (url.empty()) {
         return false;
